[BUGFIX] #5162 Fix autologin feature to work with CMS 6.2.5+
The FE session cookie is only sent with an active session since 6.2.
Since 6.2.5 a call to the (internal) API createUserSession() does not
trigger a new session automatically, hence no FE cookie is sent and
the autologin does not work.
As a workaround we force an active session by setting dummy data.

Change-Id: Ib029cd76bff22da2100a3d008f03ebdc987a8018
Reviewed-on: http://review.typo3.org/33783
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>