[TASK] Drop the elaborate validation for the user group
On the client side, we already ensure that a user group is selected
by having the radio buttons marked as required.

On the server side, we already have filtering and fallbacks in place
to ensure that only allowed user groups can be set, and that always
a user group will be set.

The only case where the additional validation would be beneficial
is to help malicious users debug their mangled input. That's not
worth the work and added complexity.

Fixes #474
Fixes #473
Closes #472