[TASK] Make displaying the group selector non-configurable
If the group selector is displayed, it should always be required
(because optional radio buttons groups do not make sense).

In addition, it does not make sense to not display the user group
selector if there is more than one user group to select from.

So make displaying the selector automatic depending on the number
of available user groups, and have the radion buttons always required.

Fixes #476